rename and set UNSUPPORTED status

+++ SYCL/ESIMD/noinline_args_size/noinline_args_char_int_size_96.cpp
@@ -1,4 +1,4 @@
-//==- noinline_func_args_char_int_size_96.cpp  - DPC++ ESIMD on-device test ==//
+//===--- noinline_args_char_int_size_96.cpp  - DPC++ ESIMD on-device test -===//
 //
 // Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
 // See https://llvm.org/LICENSE.txt for license information.
@@ -15,7 +15,7 @@
 constexpr unsigned VL = 16;
 constexpr unsigned SIZE = 96;
 
-using a_data_t = int32_t;
+using a_data_t = int8_t;
 using b_data_t = int32_t;
 using c_data_t = int32_t;
